+++ CHANGELOG.rst
@@ -10,6 +10,8 @@ This change log uses principles from `keep a changelog <http://keepachangelog.co
 Added
 ^^^^^
 
+- Added rogue content validation check to DiskStorageBroker.pre_freeze hook
+
 
 Changed
 ^^^^^^^